requirement.txt
absl-py==1.3.0
addict==2.4.0
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
brotlipy==0.7.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
clip==1.0
cosine-annealing-warmup==2.0
cryptography==38.0.1
cycler==0.11.0
flit_core==3.6.0
frozenlist==1.3.3
fsspec==2022.11.0
ftfy==5.9
google-auth==2.15.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
idna==3.4
importlib-metadata==5.1.0
joblib==1.1.1
kiwisolver==1.4.4
lightning-utilities==0.4.2
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.3.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mmcv==1.7.0
multidict==6.0.3
numpy==1.23.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
omegaconf==2.0.5
opencv-python==4.6.0.66
packaging==22.0
pandas==1.5.2
Pillow==9.3.0
pip==22.3.1
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyDeprecate==0.3.1
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-lightning==1.4.9
pytz==2022.6
PyYAML==6.0
regex==2020.11.13
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.2
setuptools==65.5.0
six==1.16.0
sklearn==0.0.post1
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
threadpoolctl==2.2.0
torch==1.13.0
torchaudio==0.13.0
torchmetrics==0.5.0
torchvision==0.14.0
tqdm==4.59.0
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
Werkzeug==2.2.2
wheel==0.37.1
yapf==0.32.0
yarl==1.8.2
zipp==3.11.0
